API

SGtSNEpi.jl exports the following functions

SGtSNEpi.sgtsnepi

SGtSNEpi.show_embedding

SGtSNEpi.sgtsnepi — Method

sgtsnepi( A::AbstractMatrix )
sgtsnepi( G::AbstractGraph )

Call SG-t-SNE-Π on the input graph, given as either a sparse adjacency matrix $A$ or a graph object $G$. Alternatively, the input can be a point-cloud data set $X$ (coordinates) of size $N \times D$, i.e.,

sgtsnepi( X::AbstractMatrix )

Optional arguments

d=2: number of dimensions (embedding space)

λ=10: SG-t-SNE scaling factor

More options (for experts)

max_iter=1000: number of iterations

early_exag=250: number of early exageration iterations

np=0: number of processors (set to 0 to use all available cores)

Y0=nothing: initial distribution in embedding space (randomly selected if nothing)

h=0: grid side length (0 for default values)

eta=200.0: learning parameter

bound_box=Inf: maximum bounding box (rescaling after that)

profile=false: disable/enable profiling. If enabled the function return a 3-tuple: (Y, t, g), where Y is the embedding coordinates, t are the execution times per iteration and g is the grid size per iteration.

Special options for point-cloud data embedding

u=10: perplexity

k=3*u: number of nearest neighbors (for kNN formation)

knn_type=( size(A,1) < 10_000 ) ? :exact : :flann: Exact or approximate kNN

Notes

Isolated are placed randomly on the top-right corner of the embedding space

The function tries to automatically detect whether the input matrix represents an adjacency matrix or data coordinates. In ambiquous cases, such as a square matrix of data coordinates, the user may specify the type using the optional argument type

:graph: the input is an adjacency matrix

:coord: the input is the data coordinates

Examples

julia> using LightGraphs

julia> Y0 = repeat(1:1000,1,2) / 1000.0;

julia> G = circular_ladder_graph( 500 )
{1000, 1500} undirected simple Int64 graph

julia> Y = sgtsnepi( G; Y0 = Y0, np = 4, early_exag = 100, max_iter = 250 );
Number of vertices: 1000
Embedding dimensions: 2
Rescaling parameter λ: 10
Early exag. multiplier α: 12
Maximum iterations: 250
Early exag. iterations: 100
Learning rate: 200
Box side length h: 0.7
Drop edges originating from leaf nodes? 0
Number of processes: 4
1000 out of 1000 nodes already stochastic
m = 1000 | n = 1000 | nnz = 3000
Working with double precision
Iteration 1: error is 96.9204
Iteration 50: error is 84.9181 (50 iterations in 0.039296 seconds)
Iteration 100: error is 4.32754 (50 iterations in 0.038005 seconds)
Iteration 150: error is 2.54655 (50 iterations in 0.066491 seconds)
Iteration 200: error is 1.90124 (50 iterations in 0.159556 seconds)
Iteration 249: error is 1.65057 (50 iterations in 0.213149 seconds)
 --- Time spent in each module ---

 Attractive forces: 0.006199 sec [1.24082%] |  Repulsive forces: 0.49339 sec [98.7592%]

SGtSNEpi.show_embedding — Function

show_embedding( Y [, L] )

Visualization of 2D embedding coordinates $Y$ using Makie. If the $n \times 1$ vector $L$ of vertex memberships is provided, points are colored accroding to the labels.

Optional arguments (for experts)

A=nothing: adjacency matrix; if provided, edges are shown

cmap: the colormap to use (default to distinguishable colors)

res=(800,800): figure resolution

lwd_in=0.5: line width for internal edges

lwd_out=0.3: line width for external edges

edge_alpha=0.2: the alpha channel for the edges

clr_out=colorant"#aabbbbbb": the color of inter-cluster edges

mrk_size=4: marker size

size_label:24: legend label size